minor changes in code plus additional unit test
