+++ examples/automl_example/custom_strategy/multimodal/multimodal.py
@@ -1,4 +1,4 @@
-from examples.example_utils import create_feature_generator_strategy
+from fedot_ind.tools.example_utils import create_feature_generator_strategy
 from fedot_ind.core.architecture.pipelines.abstract_pipeline import ApiTemplate
 from fedot_ind.core.repository.config_repository import DEFAULT_COMPUTE_CONFIG, \
     DEFAULT_AUTOML_LEARNING_CONFIG, DEFAULT_CLF_AUTOML_CONFIG
