+++ examples/automl_example/custom_strategy/specific_strategy/probability_calibration_example.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from examples.example_utils import create_feature_generator_strategy
+from fedot_ind.tools.example_utils import create_feature_generator_strategy
 from fedot_ind.core.architecture.pipelines.abstract_pipeline import ApiTemplate
 from fedot_ind.core.repository.config_repository import DEFAULT_COMPUTE_CONFIG, \
     DEFAULT_AUTOML_LEARNING_CONFIG
